Add new component called CatalogItem.tsx
